Task #1 | Refactoring Fix
Added classes:
- InputConverter;

Refactored classes:
- InputManager;
- Validator;
- Program;

Comment:
- Methods have been moved to suitable classes.
