+++ Task_1_Console_Calculator/Task_1_Console_Calculator/Task_1_Console_Calculator.csproj
@@ -48,6 +48,7 @@
     <Compile Include="MathOperations.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="InputConverter.cs" />
     <Compile Include="Validator.cs" />
   </ItemGroup>
   <ItemGroup>
